Create README.md
The world's population is now growing exponentially. It is estimated that there are about 4.2 billion people in the world, 55% of whom live in cities where the rich and the poor live together, leading to a dramatic increase in levels of inequality, as the better-off populations tend to have a higher average economic status, with better education and better health than the poor. This difference in inequality is increasingly observable in large metropolitan areas. As urbanisation and social inequality are two main policy themes of this era, reducing inequality is at the forefront of the global sustainable development agenda and is a main policy objective in many cities (GLA, 2017, 2018). However, tracking and monitoring social, economic, environmental and health outcomes in cities at high spatial and temporal resolution is therefore difficult, as the datasets to inform these policies and measure actual impacts are currently derived from disjointed and not efficient monitoring systems. Measuring socio-economic status (SES) at high spatial and temporal resolution therefore becomes an important challenge for informing and evaluating policy. In most countries and cities, such rich data are scarce and therefore measuring these inequalities is a critical issue as data on different environmental, social and health measures are often collected using very expensive processes and come from a variety of sources with different spatial resolutions and frequencies. Therefore, the use of other emerging sources of large-scale data, such as street imagery, GPS trajectories coupled with advances in deep learning methods have the potential to significantly advance and improve the speed, frequency, and proximity with which we can measure the characteristics of cities and populations in order to inform and better assess policies regarding urban inequalities. One example that is attracting the attention of some researchers is the use of street-level imagery to estimate or predict economic factors such as household income, education, crime, poverty, employment levels, unemployment, safety, housing prices, living standards etc. in a wider context. These new methods can facilitate the identification of specific areas of concern at early stages of increasing need so that interventions can be rapidly implemented through policies that aim to address them. In our work we use street-level images within Los Angeles to predict household income for small statistical areas. We compare the performance of three different models using cross-validation methods for our deep models.